정리: 레이아웃 구조 정돈
