+++ 03-native-expo/homework/web/src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import localFont from "next/font/local";
 import "./globals.css";
 
-import PageLayout from "@/common/layout";
+import PageLayout from "@/common/layout/layout";
 
 const pretendard = localFont({
     src: "../styled-system/fonts/PretendardVariable.woff2",
