+++ 03-native-expo/homework/web/src/app/place/new/page.tsx
@@ -1,67 +1,36 @@
+import ImageBox from "@/components/image-box";
 import { css } from "@/styled-system/css";
-import { Plus } from "lucide-react";
 
 export default function PlaceNewPage() {
     return (
         <>
-            <section className={css_section}>
-                <div className={css_imgBox}>
-                    <Plus size={20} />
-                    사진 등록
-                </div>
+            <ImageBox />
 
-                <label className={css_label}>
-                    플레이스 이름 <b>*</b>
-                    <input
-                        className={css(css_input)}
-                        placeholder="플레이스 이름을 입력해 주세요. (1글자 이상)"
-                    />
-                </label>
+            <label className={css_label}>
+                플레이스 이름 <b>*</b>
+                <input
+                    className={css(css_input)}
+                    placeholder="플레이스 이름을 입력해 주세요. (1글자 이상)"
+                />
+            </label>
 
-                <label className={css_label}>
-                    플레이스 주소 <b>*</b>
-                    <input className={css(css_input)} placeholder="플레이스 주소 입력" disabled />
-                </label>
+            <label className={css_label}>
+                플레이스 주소 <b>*</b>
+                <input className={css(css_input)} placeholder="플레이스 주소 입력" disabled />
+            </label>
 
-                <label className={css_label}>
-                    플레이스 내용 <b>*</b>
-                    <textarea
-                        className={css(css_input, css_text)}
-                        placeholder="플레이스 내용을 입력해 주세요. (1글자 이상)"
-                    />
-                </label>
+            <label className={css_label}>
+                플레이스 내용 <b>*</b>
+                <textarea
+                    className={css(css_input, css_text)}
+                    placeholder="플레이스 내용을 입력해 주세요. (1글자 이상)"
+                />
+            </label>
 
-                <button className={css_button}>로그 등록</button>
-            </section>
+            <button className={css_button}>로그 등록</button>
         </>
     );
 }
-
-const css_section = css({
-    w: "100%",
-    h: "100%",
-
-    display: "flex",
-    flexDir: "column",
-    gap: "2rem",
-});
-
-const css_imgBox = css({
-    w: "10rem",
-    h: "10rem",
-    bg: "#f5f5f5",
-    rounded: "0.8rem",
-    color: "#767676",
-    fontSize: "1.2rem",
-    fontWeight: "600",
-    mt: "2.4rem",
-
-    display: "flex",
-    flexDir: "column",
-    justifyContent: "center",
-    alignItems: "center",
-    gap: "0.4rem",
-});
 
 const css_label = css({
     display: "flex",
