+++ 03-native-expo/homework/web/src/common/layout/layout.tsx
@@ -4,8 +4,11 @@ import { HeaderGlobal } from "./header/header";
 export default function PageLayout({ children }: Readonly<{ children: React.ReactNode }>) {
     return (
         <div className={app}>
-            <HeaderGlobal />
-            {children}
+            <header className={css_header}>
+                <HeaderGlobal />
+            </header>
+
+            <section className={css_section}>{children}</section>
         </div>
     );
 }
@@ -23,4 +26,18 @@ const app = css({
 
     p: "2rem",
     boxShadow: "0px 4px 8px #bbb",
+});
+
+const css_header = css({
+    w: "100%",
+    h: "100%",
+});
+
+const css_section = css({
+    w: "100%",
+    h: "100%",
+
+    display: "flex",
+    flexDir: "column",
+    gap: "2rem",
 });
