+++ 03-native-expo/homework/web/src/components/input/index.tsx
@@ -1,0 +1,7 @@
+import { css } from "@/styled-system/css";
+
+export function InputBase() {
+    return (
+        <input className={css()} placeholder={} type={} autoComplete={} {...register(keyname)} />
+    );
+}
